apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
structlog==24.2.0
boto3==1.34.131
redis==5.0.7
numpy==1.26.4
pandas==2.2.2
plotly==5.24.1
scikit-learn==1.5.1
joblib==1.4.2
prophet==1.1.5
cmdstanpy==1.2.4
httpx==0.27.0
Pillow==10.4.0
matplotlib==3.9.0
jinja2==3.1.4
openpyxl==3.1.5
reportlab==4.2.2
pytest==8.3.4
ruff==0.15.9
